Refactor NullRouteImpl out of async client
