+++ source/extensions/early_data/default_early_data_policy.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <memory>
 
 #include "envoy/extensions/early_data/v3/default_early_data_policy.pb.h"
