+++ test/common/http/async_client_impl_test.cc
@@ -1870,20 +1870,20 @@ TEST_F(AsyncClientImplTest, DumpState) {
 // Must not be in anonymous namespace for friend to work.
 class AsyncClientImplUnitTest : public AsyncClientImplTest {
 public:
-  std::unique_ptr<AsyncStreamImpl::RouteImpl> route_impl_{new AsyncStreamImpl::RouteImpl(
-      client_, absl::nullopt,
+  std::unique_ptr<NullRouteImpl> route_impl_{new NullRouteImpl(
+      client_.cluster_->name(), client_.singleton_manager_, absl::nullopt,
       Protobuf::RepeatedPtrField<envoy::config::route::v3::RouteAction::HashPolicy>(),
       absl::nullopt)};
-  AsyncStreamImpl::NullVirtualHost vhost_;
-  AsyncStreamImpl::NullCommonConfig config_;
+  NullVirtualHost vhost_;
+  NullCommonConfig config_;
 
   void setupRouteImpl(const std::string& yaml_config) {
     envoy::config::route::v3::RetryPolicy retry_policy;
 
     TestUtility::loadFromYaml(yaml_config, retry_policy);
 
-    route_impl_ = std::make_unique<AsyncStreamImpl::RouteImpl>(
-        client_, absl::nullopt,
+    route_impl_ = std::make_unique<NullRouteImpl>(
+        client_.cluster_->name(), client_.singleton_manager_, absl::nullopt,
         Protobuf::RepeatedPtrField<envoy::config::route::v3::RouteAction::HashPolicy>(),
         std::move(retry_policy));
   }
